refactor: 페이지 파일 분리
ohgus <anakin1108@naver.com>
